Added changes related to migration of plugin for grails 3.x versions. And changed version and group id as per grails plugins conventions.
