--- grails-app/conf/Config.groovy
@@ -1,4 +1,0 @@
-log4j = {
-	error 'org.codehaus.groovy.grails',
-	      'org.springframework'
-}
