+++ grails-app/conf/application.groovy
@@ -1,0 +1,35 @@
+log4j = {
+	error 'org.codehaus.groovy.grails',
+	      'org.springframework'
+}
+
+springsession {
+	maxInactiveInterval = 1800
+	redis {
+		connectionFactory {
+			hostName = "localhost"
+			port = 6379
+			timeout = 2000
+			usePool = true
+			dbIndex = 0
+			convertPipelineAndTxResults = true
+		}
+		poolConfig {
+			maxTotal = 8
+			maxIdle = 8
+			minIdle = 0
+		}
+		sentinel {
+			master = null
+			nodes = []
+			password = ''
+			timeout = 2000
+		}
+	}
+	strategy {
+		defaultStrategy = "COOKIE"
+		cookie.name = "SESSION"
+		httpHeader.headerName = "x-auth-token"
+	}
+	allow.persist.mutable = false
+}
