+++ src/integration-test/groovy/ConfigurationTest.groovy
