+++ src/main/java/org/grails/plugins/springsession/converters/JdkDeserializer.java
@@ -1,4 +1,4 @@
-package grails.plugin.springsession.converters;
+package org.grails.plugins.springsession.converters;
 
 import org.springframework.core.serializer.Deserializer;
