+++ src/main/java/org/grails/plugins/springsession/data/redis/config/NoOpConfigureRedisAction.java
@@ -1,4 +1,4 @@
-package grails.plugin.springsession.data.redis.config;
+package org.grails.plugins.springsession.data.redis.config;
 
 import org.springframework.data.redis.connection.RedisConnection;
 import org.springframework.session.data.redis.config.ConfigureRedisAction;
