+++ src/main/java/org/grails/plugins/springsession/web/http/config/SpringSessionConfig.java
@@ -1,0 +1,7 @@
+package org.grails.plugins.springsession.web.http.config;
+
+import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
+
+@EnableRedisHttpSession
+public class SpringSessionConfig {
+}
